Lesson15rev1

+++ src/main/java/org/rd/lessons/lesson15/Odev/builderOdev/Urun.java
@@ -34,7 +34,7 @@ public class Urun {
             return this;
         }
 
-        public Olusturucu fiyatAyarla(int fiyat) {
+        public Olusturucu fiyatAyarla(double fiyat) {
             urun.fiyat = fiyat;
             return this;
         }
